+++ torchmetrics/classification/specificity.py
@@ -18,7 +18,6 @@ from torch import Tensor
 
 from torchmetrics.classification.stat_scores import StatScores
 from torchmetrics.functional.classification.specificity import _specificity_compute
-from torchmetrics.utilities import _deprecation_warn_arg_is_multiclass, _deprecation_warn_arg_multilabel
 
 
 class Specificity(StatScores):
@@ -105,12 +104,6 @@ class Specificity(StatScores):
         dist_sync_fn:
             Callback that performs the allgather operation on the metric state. When ``None``, DDP
             will be used to perform the allgather.
-        multilabel:
-            .. deprecated:: 0.3
-                Argument will not have any effect and will be removed in v0.4, please use ``multiclass`` intead.
-        is_multiclass:
-            .. deprecated:: 0.3
-                Argument will not have any effect and will be removed in v0.4, please use ``multiclass`` intead.
 
     Raises:
         ValueError:
@@ -142,11 +135,7 @@ class Specificity(StatScores):
         dist_sync_on_step: bool = False,
         process_group: Optional[Any] = None,
         dist_sync_fn: Callable = None,
-        multilabel: Optional[bool] = None,  # todo: deprecated, remove in v0.4
-        is_multiclass: Optional[bool] = None,  # todo: deprecated, remove in v0.4
     ):
-        _deprecation_warn_arg_multilabel(multilabel)
-        multiclass = _deprecation_warn_arg_is_multiclass(is_multiclass, multiclass)
 
         allowed_average = ["micro", "macro", "weighted", "samples", "none", None]
         if average not in allowed_average: